Converted object browser to TypeScript
- Added node interfaces
- Refresh impacted nodes instead of whole explorer
- Added progress for move/rename/delete object operation
