+++ src/locale/ids/da.ts
@@ -116,7 +116,6 @@ export const da: Locale = {
   'objectBrowser.updateMemberText.prompt': `Opdater beskrivelse af member {0}`,
   'objectBrowser.updateMemberText.errorMessage': `Fejl ved ændring af member beskrivelse! {0}`,
   'objectBrowser.renameMember.prompt': `Omdøb member {0}`,
-  'objectBrowser.renameMember.refreshMessage': `Omdøbte member. Opfrisk object browser.`,
   'objectBrowser.renameMember.errorMessage': `Fejl ved omdøbning af member! {0}`,
   'objectBrowser.uploadAndReplaceMemberAsFile.infoMessage': `Fil blev sendt til member.`,
   'objectBrowser.uploadAndReplaceMemberAsFile.errorMessage': `Fejl ved afsendelse af fil! {0}`,
@@ -153,7 +152,6 @@ export const da: Locale = {
   'objectBrowser.changeObjectDesc.prompt': `Opdater objekt beskrivelse for {0}, *BLANK for ingen beskrivelse`,
   'objectBrowser.changeObjectDesc.errorMessage': `Objekt beskrivelse må være 50 tegn eller mindre.`,
   'objectBrowser.changeObjectDesc.infoMessage': `Ændrede objekt beskrivelsen for {0} *{1}.`,
-  'objectBrowser.changeObjectDesc.infoMessage2': `Ændrede objekt beskrivelse. Opfrisk object browser.`,
   'objectBrowser.changeObjectDesc.errorMessage2': `Fejl ved ændring af objekt beskrivelse for {0}! {1}`,
   'objectBrowser.copyObject.prompt': `Kopier objekt til new bibliotek/objekt`,
   'objectBrowser.copyObject.errorMessage': `Forkert sti: {0}. Brug formatet LIB/OBJ`,
@@ -168,7 +166,6 @@ export const da: Locale = {
   'objectBrowser.renameObject.prompt': `Omdøb objekt`,
   'objectBrowser.renameObject.errorMessage': `Objektnavn må være 10 tegn eller mindre.`,
   'objectBrowser.renameObject.infoMessage': `Omdøbte objekt {0} *{1} til {2}.`,
-  'objectBrowser.renameObject.infoMessage2': `Omdøbte objekt {0} *{1} til {2}. Opfrisk object browser.`,
   'objectBrowser.renameObject.errorMessage2': `Fejl ved omdøbning af objekt {0}! {1}`,
   'objectBrowser.moveObject.prompt': `Flyt objekt`,
   'objectBrowser.moveObject.errorMessage': `Objekt navn må være 10 tegn eller mindre.`,
