+++ src/locale/ids/en.ts
@@ -29,6 +29,11 @@ export const en: Locale = {
   'sandbox.connected.modal.detail': `You are using this system at your own risk. Do not share any sensitive or private information.`,
   'sandbox.noconnection.modal.title': `Oh no! The sandbox is down.`,
   'sandbox.noconnection.modal.detail': `Sorry, but the sandbox is offline right now. Try again another time.`,
+  'not.connected':'Not connected to an IBM i',
+  'text':'Text',
+  'lines':'Lines',
+  'changed':'Changed',
+  'created':'Created',
   // ConnectionBrowser:
   'connectionBrowser.connectTo.lastConnection': `Last connection`,
   'connectionBrowser.connectTo.lastUsed': `Last used: {0}`,
@@ -116,8 +121,8 @@ export const en: Locale = {
   'objectBrowser.updateMemberText.prompt': `Update {0} text`,
   'objectBrowser.updateMemberText.errorMessage': `Error changing member text! {0}`,
   'objectBrowser.renameMember.prompt': `Rename {0}`,
-  'objectBrowser.renameMember.refreshMessage': `Renamed member. Refresh object browser.`,
   'objectBrowser.renameMember.errorMessage': `Error renaming member! {0}`,
+  'objectBrowser.renameMember.invalid.input': `New member name must be different from it's current name`,
   'objectBrowser.uploadAndReplaceMemberAsFile.infoMessage': `Member was uploaded.`,
   'objectBrowser.uploadAndReplaceMemberAsFile.errorMessage': `Error uploading content to member! {0}`,
   'objectBrowser.downloadMemberContent.infoMessage': `Member was downloaded.`,
@@ -142,6 +147,7 @@ export const en: Locale = {
   'objectBrowser.searchSourceFile.noMembers': `No members to search.`,
   'objectBrowser.searchSourceFile.errorMessage': `Error searching source members: {0}`,
   'objectBrowser.searchSourceFile.errorMessage2': `Cannot search listings using *ALL.`,
+  'objectBrowser.createFilter': `Create new filter`,
   'objectBrowser.createLibrary.prompt': `Name of new library`,
   'objectBrowser.createLibrary.errorMessage': `Cannot create library "{0}": {1}`,
   'objectBrowser.createLibrary.infoMessage': `Would you like to add the new library to the library list?`,
@@ -153,7 +159,6 @@ export const en: Locale = {
   'objectBrowser.changeObjectDesc.prompt': `Change object description for {0}, *BLANK for no description`,
   'objectBrowser.changeObjectDesc.errorMessage': `Object description must be 50 chars or less.`,
   'objectBrowser.changeObjectDesc.infoMessage': `Changed object description for {0} *{1}.`,
-  'objectBrowser.changeObjectDesc.infoMessage2': `Changed object description. Refresh object browser.`,
   'objectBrowser.changeObjectDesc.errorMessage2': `Error changing description for {0}! {1}`,
   'objectBrowser.copyObject.prompt': `Create duplicate object to new library/object`,
   'objectBrowser.copyObject.errorMessage': `Invalid path: {0}. Use format LIB/OBJ`,
@@ -165,16 +170,18 @@ export const en: Locale = {
   'objectBrowser.deleteObject.warningMessage': `Are you sure you want to delete {0} *{1}?`,
   'objectBrowser.deleteObject.infoMessage': `Deleted {0} *{1}.`,
   'objectBrowser.deleteObject.errorMessage': `Error deleting object! {0}`,
+  'objectBrowser.deleteObject.progress':'Deleting object {0}...',
   'objectBrowser.renameObject.prompt': `Rename object`,
   'objectBrowser.renameObject.errorMessage': `Object name must be 10 chars or less.`,
   'objectBrowser.renameObject.infoMessage': `Renamed object {0} *{1} to {2}.`,
-  'objectBrowser.renameObject.infoMessage2': `Renamed object {0} *{1} to {2}. Refresh object browser.`,
   'objectBrowser.renameObject.errorMessage2': `Error renaming object {0}! {1}`,
+  'objectBrowser.renameObject.progress': `Renaming object {0} *{1} to {2}...`,
   'objectBrowser.moveObject.prompt': `Move object`,
   'objectBrowser.moveObject.errorMessage': `Object name must be 10 chars or less.`,
   'objectBrowser.moveObject.infoMessage': `Moved object {0} *{1} to {2}.`,
   'objectBrowser.moveObject.infoMessage2': `Moved object {0} *{1} to {2}. Refresh object browser.`,
   'objectBrowser.moveObject.errorMessage2': `Error moving object {0}! {1}`,
+  'objectBrowser.moveObject.progress': `Moving object {0} *{1} to {2}...`,
   // ProfilesView:
   'ProfilesView.saveConnectionProfile.prompt': `Name of profile`,
   'ProfilesView.saveConnectionProfile.infoMessage': `Saved current settings to profile '{0}'.`,
