+++ src/locale/ids/fr.ts
@@ -29,6 +29,11 @@ export const fr: Locale = {
   'sandbox.connected.modal.detail': `Vous êtes seul responsable de l'utilisation du bas à sable. Ne partagez pas d'informations à caractère sensible ou privé.`,
   'sandbox.noconnection.modal.title': `Aïe aïe aïe, le bac à sable est en panne!`,
   'sandbox.noconnection.modal.detail': `Désolé, le bac à sable est est hors-ligne pour le moment. Réessayez plus tard.`,
+  'not.connected':'Non connecté à un IBM i',
+  'text':'Texte',
+  'lines':'Lignes',
+  'changed':'Modification',
+  'created':'Création',
   // ConnectionBrowser:
   'connectionBrowser.connectTo.lastConnection': `Dernière connexion`,
   'connectionBrowser.connectTo.lastUsed': `Dernière utilisation: {0}`,
@@ -112,8 +117,8 @@ export const fr: Locale = {
   'objectBrowser.updateMemberText.prompt': `Mise à jour du texte de {0}`,
   'objectBrowser.updateMemberText.errorMessage': `Erreur lors de la mise à jour du texte du membre! {0}`,
   'objectBrowser.renameMember.prompt': `Renommer {0}`,
-  'objectBrowser.renameMember.refreshMessage': `Membre renommé. Rafraîchissement de l'explorateur d'objet.`,
   'objectBrowser.renameMember.errorMessage': `Erreur lors du renommage du membre! {0}`,
+  'objectBrowser.renameMember.invalid.input': `Le nouveau nom du membre doit être différentd de l'ancien`,
   'objectBrowser.uploadAndReplaceMemberAsFile.infoMessage': `Le membre a été uploadé.`,
   'objectBrowser.uploadAndReplaceMemberAsFile.errorMessage': `Erreur lors de l'upload du contenu du membre! {0}`,
   'objectBrowser.downloadMemberContent.infoMessage': `Le membre a été téléchargé.`,
@@ -138,6 +143,7 @@ export const fr: Locale = {
   'objectBrowser.searchSourceFile.noMembers': `Aucun membre à rechercher.`,
   'objectBrowser.searchSourceFile.errorMessage': `Erreur lors de la recherche de membres source: {0}`,
   'objectBrowser.searchSourceFile.errorMessage2': `Impossible de rechercher une liste avec *ALL.`,
+  'objectBrowser.createFilter': `Créer un filtre`,
   'objectBrowser.createLibrary.prompt': `Nom de la nouvelle bibliothèque`,
   'objectBrowser.createLibrary.errorMessage': `Impossible de créer la bibliothèque "{0}": {1}`,
   'objectBrowser.createLibrary.infoMessage': `Voulez-vous ajouter la nouvelle bibliothèque à la liste des bibliothèques?`,
@@ -149,7 +155,6 @@ export const fr: Locale = {
   'objectBrowser.changeObjectDesc.prompt': `Changer la description de l'objet pour {0}, *BLANK pour effacer la description`,
   'objectBrowser.changeObjectDesc.errorMessage': `La description de l'objet doit faire au maximum 50 caractères de long.`,
   'objectBrowser.changeObjectDesc.infoMessage': `La description de l'objet {0} *{1} a été changée.`,
-  'objectBrowser.changeObjectDesc.infoMessage2': `Description de l'objet changée. Rafraîchissement de l'explorateur d'objet.`,
   'objectBrowser.changeObjectDesc.errorMessage2': `Erreur lors du changement de la description de {0}! {1}`,
   'objectBrowser.copyObject.prompt': `Duplication de l'objet vers une nouvelle bibliothèque/objet`,
   'objectBrowser.copyObject.errorMessage': `Chemin invalide: {0}. Utilisez le format LIB/OBJ`,
@@ -161,16 +166,18 @@ export const fr: Locale = {
   'objectBrowser.deleteObject.warningMessage': `Êtes-vous sûr de vouloir supprimer {0} *{1}?`,
   'objectBrowser.deleteObject.infoMessage': `{0} *{1} supprimé.`,
   'objectBrowser.deleteObject.errorMessage': `Erreur lors de la suppression de l'objet! {0}`,
+  'objectBrowser.deleteObject.progress':`Suppression de l'objet {0}...`,
   'objectBrowser.renameObject.prompt': `Renamer l'objet`,
   'objectBrowser.renameObject.errorMessage': `Le nom de l'objet doit faire au maximum 10 caractères de long.`,
   'objectBrowser.renameObject.infoMessage': `Objet {0} *{1} renommé {2}.`,
-  'objectBrowser.renameObject.infoMessage2': `Objet {0} *{1} renommé {2}. Rafraîchissement de l'explorateur d'objet.`,
+  'objectBrowser.renameObject.progress': `Renommage de l'objet {0} *{1} en {2}...`,
   'objectBrowser.renameObject.errorMessage2': `Erreur lors du renommage de l'objet {0}! {1}`,
   'objectBrowser.moveObject.prompt': `Déplacer l'objet`,
   'objectBrowser.moveObject.errorMessage': `Le nom de l'objet doit faire au maximum 10 caractères de long.`,
   'objectBrowser.moveObject.infoMessage': `Objet {0} *{1} déplacé vers {2}.`,
   'objectBrowser.moveObject.infoMessage2': `Objet {0} *{1} déplacé vers {2}. Rafraîchissement de l'explorateur d'objet.`,
   'objectBrowser.moveObject.errorMessage2': `Erreur lors du déplacement de l'objet {0}! {1}`,
+  'objectBrowser.moveObject.progress': `Déplacement de l'objet {0} *{1} vers {2}...`,
   // ProfilesView:
   'ProfilesView.saveConnectionProfile.prompt': `Nom du profil`,
   'ProfilesView.saveConnectionProfile.infoMessage': `Paramètres actuels sauvés sur le profil '{0}'.`,
